requirements.txt
Flask==1.1.2
gunicorn==20.0.4
flask-pymongo==2.2.0
dnspython==1.16.0
python-dotenv==0.13.0
bcrypt==5.0.0